video-7-local-rag-integration/requirements.txt
sentence-transformers>=2.0.0,<3.0.0
faiss-cpu
numpy
